+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/featureRegistry/ewgf/EwgfDiscordFeature.kt
@@ -54,6 +54,9 @@ internal class EwgfDiscordFeature(
         operation: EwgfOperations.Operation,
     ): Result<BotOutput, BotError> {
         val result = when (operation) {
+            is EwgfOperations.Operation.Help -> {
+                Result.Success(ewgfHelpEmbed(featureInfo))
+            }
             is EwgfOperations.Operation.Register -> {
                 registerPlayerUseCase.invoke(
                     discordId = discordId,
@@ -76,6 +79,11 @@ internal class EwgfDiscordFeature(
             is EwgfOperations.Operation.Unregister -> {
                 unregisterPlayerUseCase.invoke(discordId)
                     .map { successEmbed(operation, featureInfo) }
+            }
+            is EwgfOperations.Operation.Search -> {
+                getPlayerUseCase.invoke(operation.discordId).map { data ->
+                    recentSetsEmbed(data, featureInfo)
+                }
             }
         }