+++ bot/discord/src/jvmTest/kotlin/io/github/sophon/botdiscord/featureRegistry/ewgf/usecase/ParseQueryIntoOperationUseCaseTest.kt
@@ -1,0 +1,56 @@
+package io.github.sophon.botdiscord.featureRegistry.ewgf.usecase
+
+import assertk.assertThat
+import assertk.assertions.isEqualTo
+import assertk.assertions.isInstanceOf
+import io.github.sophon.discord.featureRegistry.ewgf.EwgfOperations
+import io.github.sophon.discord.featureRegistry.ewgf.usecase.ParseQueryIntoOperationUseCase
+import kotlin.test.Test
+import io.github.sophon.core.domain.Result
+
+class ParseQueryIntoOperationUseCaseTest {
+    private val usecase = ParseQueryIntoOperationUseCase()
+
+    @Test
+    fun `usecase recognizes no parameters as data`() {
+        // given
+        val query = ""
+        val expectedResult = Result.Success::class
+        val expectedOp = EwgfOperations.Operation.Data
+
+        // when
+        val result = usecase.invoke(query)
+
+        //then
+        assertThat(result).isInstanceOf(expectedResult)
+        assertThat((result as Result.Success).data).isEqualTo(expectedOp)
+    }
+
+    @Test
+    fun `usecase handles tag`() {
+        // given
+        val query = "<@786351781168939038>"
+        val expectedResult = Result.Success::class
+        val expectedOp = EwgfOperations.Operation.Search("786351781168939038")
+
+        // when
+        val result = usecase.invoke(query)
+
+        //then
+        assertThat(result).isInstanceOf(expectedResult)
+        assertThat((result as Result.Success).data).isEqualTo(expectedOp)
+    }
+
+    @Test
+    fun `usecase detects bad tag`() {
+        // given
+        val query = "@111111111111"
+        val expectedResult = Result.Error::class
+
+        // when
+        val result = usecase.invoke(query)
+
+        //then
+        assertThat(result).isInstanceOf(expectedResult)
+    }
+}