+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # BOT CHANGELOG
 
+## [v11.1.0] - 2026-03-27
+- `SuperCombo`
+  - add `cr` motion input support
+  - fixed `360` inputs requiring `+`
+- `DustLoop`
+  - split move properties to basic and detailed
+- `ewgf`
+  - fixed slash command not showing the set list
+  - `help` (alt input `?`) operation
+  - `search` operation
+
 ## [v11.0.0] - 2026-03-25
 - `stats` - show the total bot usage count
 - `SuperCombo`